Added --precise-progress option to mkvmerge

+++ src/input/r_matroska.h
@@ -208,7 +208,7 @@ public:
   virtual void read_headers();
   virtual file_status_e read(generic_packetizer_c *ptzr, bool force = false);
 
-  virtual int get_progress();
+  virtual float get_progress();
   virtual void set_headers();
   virtual void identify();
   virtual void create_packetizers();
