+++ src/input/r_real.h
@@ -101,7 +101,7 @@ public:
 
   virtual void read_headers();
   virtual file_status_e read(generic_packetizer_c *ptzr, bool force = false);
-  virtual int get_progress();
+  virtual float get_progress();
   virtual void identify();
   virtual void create_packetizers();
   virtual void create_packetizer(int64_t tid);
